refactor: split distributed_tooling generate.rs into focused modules
generate/ becomes a module: mod.rs (Scaffold + orchestration + entry + tests),
names.rs (name/message normalization + validation), service_crate.rs (the Rust
templates as impl Scaffold), github.rs (repo parsing). Sets up gitops.rs and the
GitHub workflow templates as their own files for the next slice. No behavior
change; 7 tests green.

Implements [[specs/distributed-service-scaffold-tooling]]
